+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/asbuilt/repository/AssetAsBuildSpecification.java
@@ -23,11 +23,18 @@ import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
+import jakarta.persistence.criteria.Subquery;
 import org.eclipse.tractusx.traceability.assets.infrastructure.asbuilt.model.AssetAsBuiltEntity;
 import org.eclipse.tractusx.traceability.common.model.SearchCriteriaFilter;
+import org.eclipse.tractusx.traceability.common.model.SearchCriteriaStrategy;
 import org.eclipse.tractusx.traceability.common.repository.BaseSpecification;
+import org.eclipse.tractusx.traceability.qualitynotification.domain.base.model.QualityNotificationStatus;
+import org.eclipse.tractusx.traceability.qualitynotification.infrastructure.alert.model.AlertEntity;
+import org.eclipse.tractusx.traceability.qualitynotification.infrastructure.investigation.model.InvestigationEntity;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.data.jpa.domain.Specification;
+
+import java.util.List;
 
 
 public class AssetAsBuildSpecification extends BaseSpecification<AssetAsBuiltEntity> implements Specification<AssetAsBuiltEntity> {
@@ -37,6 +44,49 @@ public class AssetAsBuildSpecification extends BaseSpecification<AssetAsBuiltEnt
 
     @Override
     public Predicate toPredicate(@NotNull Root<AssetAsBuiltEntity> root, @NotNull CriteriaQuery<?> query, @NotNull CriteriaBuilder builder) {
+        if (SearchCriteriaStrategy.NOTIFICATION_COUNT_EQUAL.equals(getSearchCriteriaFilter().getStrategy())) {
+            return activeNotificationCountEqualPredicate(getSearchCriteriaFilter(), root, builder, getSearchCriteriaFilter().getValue());
+        }
+
         return createPredicate(getSearchCriteriaFilter(), root, builder);
+    }
+
+
+    private static Predicate activeNotificationCountEqualPredicate(SearchCriteriaFilter criteria, Root<?> root, CriteriaBuilder builder, String expectedFieldValue) {
+        String joinTableName;
+        Class<?> notificationClass;
+        switch (criteria.getKey()) {
+            case "activeAlertsCount" -> {
+                notificationClass = AlertEntity.class;
+                joinTableName = "alerts";
+            }
+            case "activeInvestigationsCount" -> {
+                notificationClass = InvestigationEntity.class;
+                joinTableName = "investigations";
+            }
+            default -> throw new UnsupportedOperationException();
+        }
+
+        CriteriaQuery<?> cq = builder.createQuery();
+
+        Subquery<Long> sub = cq.subquery(Long.class);
+        Root<?> subRoot = sub.from(notificationClass);
+        sub.select(builder.count(subRoot.get("id")));
+        sub.where(
+                builder.or(
+                        activeNotificationPredicates(builder, subRoot)
+                )
+        );
+        root.join(joinTableName);
+
+        return builder.equal(sub, Integer.parseInt(expectedFieldValue));
+    }
+
+    private static Predicate[] activeNotificationPredicates(CriteriaBuilder builder, Root<?> root) {
+        List<Predicate> predicates = QualityNotificationStatus.getActiveStates().stream()
+                .map(state -> builder.equal(root.get("status").as(String.class), state.name()))
+                .toList();
+
+        return predicates.toArray(new Predicate[0]);
     }
 }